refactor: move interface to consumer side
