+++ internal/usecase/maincateg/maincateg_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/eyo-chen/expense-tracker-go/internal/domain"
-	"github.com/eyo-chen/expense-tracker-go/internal/usecase/interfaces"
 	"github.com/eyo-chen/expense-tracker-go/mocks"
 	"github.com/eyo-chen/expense-tracker-go/pkg/testutil"
 	"github.com/stretchr/testify/suite"
@@ -17,7 +16,7 @@ var (
 
 type MainCategSuite struct {
 	suite.Suite
-	mainCategUC       interfaces.MainCategUC
+	mainCategUC       *MainCategUC
 	mockIconRepo      *mocks.IconRepo
 	mockMainCategRepo *mocks.MainCategRepo
 }
