+++ internal/usecase/subcateg/subcateg.go
@@ -1,8 +1,8 @@
 package subcateg
 
 import (
-	"github.com/eyo-chen/expense-tracker-go/internal/adapter/interfaces"
 	"github.com/eyo-chen/expense-tracker-go/internal/domain"
+	"github.com/eyo-chen/expense-tracker-go/internal/usecase/interfaces"
 )
 
 type SubCategUC struct {
