+++ internal/usecase/user/user.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"time"
 
-	"github.com/eyo-chen/expense-tracker-go/internal/adapter/interfaces"
 	"github.com/eyo-chen/expense-tracker-go/internal/domain"
+	"github.com/eyo-chen/expense-tracker-go/internal/usecase/interfaces"
 	"github.com/eyo-chen/expense-tracker-go/pkg/auth"
 	"github.com/eyo-chen/expense-tracker-go/pkg/logger"
 )
